Created a file storage layer API, which handles storing files (personal files for users and attachments for messages). This layer should make it very easy to implement different file storage systems by means of a module. This layer is still work in progress. There are probably a few more file related functions that can go in this file storage API.
I also started on an idea for providing a real Phorum API, by moving generic
code from the front end scripts to API libraries. The goal that I have in
mind for this is to make it possible to add additional API's (like Ajax and
SOAP) to Phorum, without having to write API code that copies loads and loads
of the core code from the front end scripts. A structured and documented API
for Phorum should also provide better and stable support for module writers.
The API code can be found in include/api/*.