Added a mechanism for allowing modules to implement functionalty of missing PHP extensions. The functions and extensions that are needed by Phorum, are hard-coded at the start of the common.php script. This script will check if all required functions are available. If not, then code from common.php will check if there is a module available which implements the missing extension. The name of that module should be: mods/compat_<extension>. The php file for the module (mods/compat_<extension>/<extension>.php) will be loaded at the start of common.php.
If no module is available to fill in the gap, then an error will be shown.
A suggestion is done to either install the extension on PHP or download
a compatibiliy module from phorum.org.

In the module list, the module will be shown and activated automatically
when the module is needed. If the module is not needed, it will not be
shown in the module page. The module system can decide on this, based
on a new field in the info.txt: "compat". The format for this field is:
"compat: <extension>|<function>".  This means: "This module implements
function <function> for extension <extension>." 

Two compatibility modules were created, based on already existing
compatibility hacks in the core code: compat_mbstring and compat_json.