Setup a more consistent language file naming scheme (we've talked about doing this ages ago, so it was about time): <languagecode>.<encoding>.php, e.g. nl_NL.UTF-8.php or nl_NL-informal.ISO-8859-1.php.
To prevent issues with modules that provide old language file names, I built
in a new language config file feature: $PHORUM['compat_languages']. This
contains an array of language names that are compatible with the language
as described by the language file. For example for the new "en_US.UTF-8"
language, we add "english" to the list of compatible languages.
This compatibility array is used by the module language loading code to
fallback to a compatible language if the main language is not available
(the first language file that is found is the winner).

Additionally, some code was added to trigger an error in case no
language file can be found.