Improved get_recent_messages() for retrieving unread messages.
Extra queries are executed to find the total number of unread messages
and the forums in which those messages are stored. This info can be used
to greatly improve the speed of the get_recent_messages() call.

The biggest improvement is when there are less unread messages than
requested in the $limit function parameter. By tuning down the limit,
the database server won't do a complete message table scan, because it
is trying to fill up the full $limit before returning data.