Dumped the "enable_new_pm_count" setting in favor of a live new pm count.
A new field "pm_new_count" was added to the user table. This field keeps
track of the number of new private messages for a user, by means of new
updates that were added to the PM functions in the DB layer.

By updating the new field when needed, there is no further need to run
a query to find the number of new messages. This works around the
performance issue for which the "enable_new_pm_count" option was
originally introduced.

A database integrity check was added to the admin interface to rebuild
all new PM counts for all users.

In the templates, {USER->new_private_messages} now contains the actual
number of new messages, not just TRUE or FALSE.