When creating a folder/forum, use the active parent_id if available.
This makes creating a folder/forum hierarchy easier for the
administrator. When browsing the hierarchy under "Manage Forums",
the links to "Create Forum" and "Create Folder" will now create
an entry inside the active folder by default (the admin can still
change the parent folder in the folder or forum form).